fix: エラー制御に使う変数を別クラスのフィールド変数を用いるようにした
