+++ src/main/java/jp/co/esm/miffy/Controller/MemberInfoController.java
@@ -1,15 +1,22 @@
 package jp.co.esm.miffy.Controller;
 
 import jp.co.esm.miffy.entity.Asf4Member;
+import jp.co.esm.miffy.form.ErrorCheck;
 import jp.co.esm.miffy.service.Asf4MemberService;
+import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import lombok.Setter;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.SessionAttributes;
 
+import java.util.List;
 import java.util.NoSuchElementException;
 
+@Setter
+@Getter
 @Controller
 @RequiredArgsConstructor
 @SessionAttributes(types = Asf4Member.class)
@@ -18,7 +25,7 @@ public class MemberInfoController {
     /**
      * 入力フォームをリセットするかどうかを判定する変数
      */
-    private boolean formReset = true;
+    private boolean formReset = false;
 
     /**
      * entityオブジェクトを初期化して返却する
@@ -31,6 +38,15 @@ public class MemberInfoController {
         return asf4Member;
     }
 
+    @RequestMapping("asf4members")
+    public String index(Model model) {
+        formReset = false;
+        List<Asf4Member> asf4MemberList = asf4MemberService.selectAll();
+        model.addAttribute("asf4MemberList", asf4MemberList);
+        // asf4MemberService.hook();    // デバッグ用のhookメソッドの呼び出し
+        return "asf4members";
+    }
+
     /**
      * 検索画面に遷移する
      * entityオブジェクト"asf4Member"のidobataIdに文字列"no"をsetすることで、検索画面でエラーメッセージが表示されないようにしている。
@@ -39,9 +55,9 @@ public class MemberInfoController {
      * @return 検索画面へのパス
      */
     @RequestMapping("/search")
-    public String search(Asf4Member asf4Member) {
+    public String search(Asf4Member asf4Member, ErrorCheck errorCheck) {
+        errorCheck.setCheck("no");
         asf4Member.setIdobataId("");
-        asf4Member.setName("no");
         return "search";
     }
 
@@ -55,7 +71,7 @@ public class MemberInfoController {
      * @return 確認画面か検索画面へのパス
      */
     @RequestMapping("/confirm")
-    public String confirm(Asf4Member asf4Member) {
+    public String confirm(Asf4Member asf4Member, ErrorCheck errorCheck) {
         try {
             Asf4Member asf4MemberOptional = asf4MemberService.selectByidobataId(asf4Member.getIdobataId());
             asf4Member.setId(asf4MemberOptional.getId());
@@ -63,10 +79,10 @@ public class MemberInfoController {
             asf4Member.setIdobataId(asf4MemberOptional.getIdobataId());
             asf4Member.setFloor(asf4MemberOptional.getFloor());
             asf4Member.setSkip(asf4MemberOptional.isSkip());
-            formReset = false;
+            formReset = true;
             return "confirm";
         } catch (NoSuchElementException e) {
-            asf4Member.setName("Yes");
+            errorCheck.setCheck("Yes");
             System.out.println("エラーをcatchしました");
             return "search";
         }
@@ -81,12 +97,10 @@ public class MemberInfoController {
      */
     @RequestMapping("/update")
     public String update(Asf4Member asf4Member) {
-        if(formReset) {
+        if (!formReset) {
             asf4Member.setName("");
             asf4Member.setIdobataId("");
             asf4Member.setFloor("");
-        } else {
-            formReset = true;
         }
         return "update";
     }
@@ -98,7 +112,7 @@ public class MemberInfoController {
      */
     @RequestMapping("/complete")
     public String complete(Asf4Member asf4Member) {
-            asf4MemberService.update(asf4Member);
-            return "complete";
+        asf4MemberService.update(asf4Member);
+        return "complete";
     }
 }
