+++ src/main/java/jp/co/esm/miffy/form/ErrorCheck.java
@@ -1,0 +1,10 @@
+package jp.co.esm.miffy.form;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ErrorCheck {
+    private String check;
+}
