+++ src/main/resources/templates/search.html
@@ -6,11 +6,12 @@
 </head>
 <body>
 <p>下記必要事項を記載の上、「検索」ボタンを押下してください。</p><br/>
-</form>
-<form method="post" action="/confirm" th:object="${asf4Member}">
-    <div th:if="*{name}">
+<form method="post" action="/confirm" th:object="${errorCheck}">
+    <div th:if="*{check}">
         <font color="red"><p th:text="検索条件にマッチする情報がありませんでした。"></p></font>
     </div>
+</form>
+<form method="post" action="/confirm" th:object="${asf4Member}">
     idobataID：<input type="text" th:value="*{idobataId}" th:field="*{idobataId}"/>
     <br/>
     <input type="submit" value="検索"/>
